requirements.txt
antlr4-python3-runtime==4.9.3
array-api-compat==1.12.0
biom_format==2.1.17
biopython==1.86
certifi==2025.10.5
cffi==1.18.0.dev0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
fonttools==4.60.1
greenlet==0.4.13
h5py==3.15.1
hpy==0.9.0
hydra-core==1.3.2
idna==3.11
iniconfig==2.3.0
kiwisolver==1.4.9
loguru==0.7.3
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
natsort==8.4.0
numpy==2.2.6
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
pluggy==1.6.0
pyCirclize==1.10.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-color-ext==0.1.8
rich-gradient==0.3.6
six==1.17.0
statsmodels==0.14.5
tomli==2.3.0
typer==0.12.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0